Remove 'in' core function.
The 'in' core function was added as a replacement for the new 'get'
semantics, however it is actually an almost entirely redundant and
confusing function.

The reason the function is redundant is because it can nearly be trivially
replaced by just indexing on the datastructure directly.

The reason the function is not named well and confusing is because the
name does not match the underlying bytecode, 'in' was implemented by the
'get' bytecode and get was implemented by the get c function, a very
confusing situation.

This pull request simply removes 'in' because it is not needed.