Rename in to ind
Why:

Consider (in t k), it could be considered by a user to be set membership
at first glance, but actually is indexing and lookup. This may be a
source of confusion for many people in the future we can avoid with some
changes now.

Notes:

Unfortunately some code in boot uses the name ind, in the cases it
conflicted the variables were renamed to indexed.

Note this change still leaves an inconsistency where the ind function
is mostly implemented by the OP_GET opcode. It's possible to rename to
OP_IND, but then we should also rename janet_get() too. It also leaves
the question about abstract types having a get method and not an ind
method... Overall those issues seem less of a concern.