Add struct special-forms to root environment
Janet contains a number of special forms which are defined in the
compiler. Currently, these forms are not accessible from "within" Janet.
Where references to the special forms are required, they must be
hard-coded. This is error-prone and can result in references becoming
stale.

This commit uses the existing `janetc_specials` array that is defined in
`src/core/compile.c` to generate a `special-forms` struct. To do this
requires exposing the `janetc_specials` array through `compile.h`. At
present, the value for each symbol is itself a struct that contains a
docstring under the `:doc` keyword.

The `doc` macro is updated to use the `special-forms` struct to display
documentation for the special forms. The wording is based on the wording
used on the section of the Janet website explaining the special forms.

This commit also cleans up other hard-coded references to instead use
either the `special-forms` struct or `janetc_specials`.